Move Pages API into Route Handlers
